refactor(user): 重构用户认证逻辑
- 将用户认证相关的逻辑从 `UserLogic` 移动到新的 `AuthLogic` 类中。
- 更新 `UserService` 接口，移除与认证相关的接口方法。
- 删除 `UserController` 中的用户登录和注册接口。
- 在 `BuildingService` 中添加 `@NotNull` 注解以增强返回值的非空约束。
- 将 `UserTest` 测试类重命名为 `AuthTest`。
